Set text color to --result-text-color
Previously, we set the text color to `--search-text-color`.
That would create contrast issues if search background
color is considerably different from the window color.

Since the background color for the HTML is set to
`--window-color`, and since in the Alfred UI, the
result text is displayed directly on top of it, we
can be sure that `--window-color` and `--result-text-color`
will always provide good contrast. (and if they don't,
then the Alfred theme is as useless as the extra pane).