Hardlink the stylesheet instead of creating a copy
To tweak the CSS, one would need to navigate to
the data dir, then to the HTML dir of the specific
dictionary, and edit `dict-entry.css`.

That flow is non-obvious. Moreover, it is confusing
because there's a `dict-entry.css` in the workflow
dir, and it could get annoying to see that editing
it isn't reflecting the changes in dictionaries that
have already been imported.

In this change, while importing a new dict, instead
of copying the css file from the workflow dir to
the data dir, we create a hard link instead. That way,
editing the css file in the workflow dir is enough.
Also, it has the added benefit of reflecting the changes
in all the imported dictionaries in one go.