Handle dicts without `CFBundleDisplayName` in their info.plist
There seem to be dictionaries whose `info.plist` does not
contain the property `CFBundleDisplayName`.

I wasn't able to reproduce the issue locally because all
the dictionaries Apple allows me to enable on my laptop
do contain that property.

Based on #9, it looks like these problematic dictionaries
do have the `CFBundleName` property. To handle such such
dictionaries, in this change, we introduce a fallback order:
 - `CFBundleDisplayName`
 - `CFBundleName`
 - `CFBundleIdentifier`

We also make sure that if a dictionary doesn't have any of
the above attributes, we ignore the dictionary, log the
error, but do not fail `.dict-import`.

Fixes #9