Discover dicts directly storing data in Contents/Body.data
While all dictionaries that come bundled with macOS
store the `Body.data` file in `Contents/Resources` folder,
https://agiletortoise.com/terminology/mac/ stores it
directly in the `Contents` folder.

With this change, we make sure to handle that case as
well.

At the moment, importing the Terminology dictionary still
leads to visual glitches in the previews, which we may fix
if future patches.