Include python "fat binary" in workflow
Instead of compiling `BetterDict.py` in a binary and
distributing it, we instead create a "fat python binary"
(like fat JAR) that mirrors the behavior of the python
binary, but comes with all the required pip packages
embedded. Along with the compiled "python" binary, we
now include the python scripts in the workflow.

The main advantage of this method is faster local development
of the workflow. Previously, the flow was:
1. Edit the python scripts
2. Run mkapps.sh
3. Manually test the workflow

Now, the step (2) isn't required.